+++ src/checkout/index.html
@@ -9,6 +9,7 @@
     <title>Sleep Outside | Checkout</title>
 
     <link rel="stylesheet" href="css/style.css" />
+    <script src="../js/checkout.js" type="module"></script>
   </head>
 
   <body>
