+++ src/js/product.js
@@ -1,6 +1,9 @@
 import { getParam } from "./utils.mjs";
 import ProductData from "./ProductData.mjs";
 import ProductDetails from "./ProductDetails.mjs";
+import { loadHeaderFooter } from "./utils.mjs";
+
+loadHeaderFooter();
 
 const productId = getParam("product");
 const dataSource = new ProductData("tents");
